Test: organize GEMM projection tests
