+++ docs/fe-oss-apis/gemm_fusions/gemm_proj_rope_mxfp8.md
@@ -149,4 +149,4 @@ pip install nvidia-cudnn-frontend[cutedsl]
 
 ## Usage examples
 
-For usage examples, see test cases in `test/python/fe_api/test_gemm_proj_rope_mxfp8.py`.
+For usage examples, see test cases in `test/python/fe_api/gemm/test_gemm_proj_rope_mxfp8.py`.
